[SW-14143] Added an ETSoC1 target-specific assembly constraint for m0
We needed a way of specifying the m0 mask register for inline assembly
that not only generated the needed code, but was also suitable for
optimizing out instructions redundantly setting the mask to exactly the
same value it already had.

The cleanest way I came up with was to implement in the RISC-V backend
a new ETSoC1 especific assembly constraint that maps an operand to the
m0 mask register.

That constraint is used for assigning a register, which is always m0,
and setting a given value passed through the initialization expression.
Now one can write code like this:

  float __attribute__((vector_size(sizeof(f32_t) * 8))) fdst;
  float __attribute__((vector_size(sizeof(f32_t) * 8))) fsrc1;
  float __attribute__((vector_size(sizeof(f32_t) * 8))) fsrc2;
  __asm__ volatile("fbci.ps %[src1], %[value1]\n"
                   "fbci.ps %[src2], %[value2]\n"
                   "fadd.ps %[dst], %[src1], %[src2]\n"
                   : [ dst ] "=f"(fdst), [ src1 ] "=f"(fsrc1),
                     [ src2 ] "=f"(fsrc2)
                   : [ value1 ] "i"(1), [ value2 ] "i"(2), [mask]
"M"(255));

  __asm__ volatile("nop\n"
                   "nop\n"
                   "nop\n"
                   "nop\n");

  static_assert(sizeof(u32_t) == 4);
  u32_t __attribute__((vector_size(sizeof(u32_t) * 8))) udst;
  u32_t __attribute__((vector_size(sizeof(u32_t) * 8))) usrc1;
  u32_t __attribute__((vector_size(sizeof(u32_t) * 8))) usrc2;
  __asm__ volatile("fbci.ps %[src1], %[value1]\n"
                   "fbci.ps %[src2], %[value2]\n"
                   "fadd.pi %[dst], %[src1], %[src2]\n"
                   ""
                   : [ dst ] "=f"(udst), [ src1 ] "=f"(usrc1),
                     [ src2 ] "=f"(usrc2)
                   : [ value1 ] "i"(1), [ value2 ] "i"(2), [mask]
"M"(255));

And the resulting code only sets the mask only once, like in the
following generated code:

0000000000000002 <main>:
   2:   0ff00513                li      a0,255
   6:   6621707b                maskand mt0,mt2,mt2
   a:   560500fb                mov.m.x mt1,a0,0x0
   e:   0000109f                fbci.ps ft1,0x1
  12:   0000211f                fbci.ps ft2,0x2
  16:   0020f07b                fadd.ps ft0,ft1,ft2
  1a:   0001                    nop
  1c:   0001                    nop
  1e:   0001                    nop
  20:   0001                    nop
  22:   0000109f                fbci.ps ft1,0x1
  26:   0000211f                fbci.ps ft2,0x2
  2a:   0620807b                fadd.pi ft0,ft1,ft2

Because the m0 register is set by the LLVM infraestructure, it
comes for free to optimize redundant instructions for setting the
mask to exactly the same value it already has.

Note that the sequence for setting the mask is using three instructions
when it could use just one. I have created a ticket for cleaning up this
inefficiency.