[SW-16819] Fix for broken debug build
A previous fix done on RISCVInstrInfoEsperanto.td did not follow a
regeneration of the replacement patterns for Esperanto's instrinsics.

Whenever changes are done on RISCVInstrInfoEsperanto.td the following
commands need to be run for regenerating the replacement patterns:

python3 esperanto/python/intrinsics.py

cp RISCVInstrInfoEsperantoPatterns.td \
   llvm/lib/Target/RISCV/RISCVInstrInfoEsperantoPatterns.td

cp BuiltinsRISCVET.def ./clang/include/clang/Basic/BuiltinsRISCVET.def

cp IntrinsicsRISCVET.td ./llvm/include/llvm/IR/IntrinsicsRISCVET.td

In fairness, the need for this regeneration step was not obvious and this
should be part of the build system, but it is what it is. In this case we
noticed the issue of the patterns being out of sync  because extra checks
are performed on debug builds.