[SW-16746] Supporting memory operands like "0(rs)" and "(rs)" for shg
The code from et-common-libs used the "A" memory operand, that in turn
was converted into the "0(rs)" instruction format, which was not
supported by the shg instruction. Our shg instruction only supported the
"(rs)" format, in line with ETSoC reference manual.

We needed a way to accept both formats, in order to compile this code.
I found out that other LLVM users had exactly the same problem when
implementing the official atomic instructions and the GPRMemAtomic
operand class was introduced for this particular case.

Our LLVM fork is old, but already had the aforementioned operand class
and we just needed to use GPRMemAtomic as a drop-in replacement for GPR.

Here is the patch that inspired the solution:
https://reviews.llvm.org/rGf596f45070a0db27c431b0585248f61a40acbb1c