[SW-16747] Custom lowering for casting between int64_t/uint64_t and float
ETSoC1 lacks the four standard RISC-V convert instructions needed for
static casting between 64 bits integers and floats:

- fcvt.s.lu: for converting an unsigned 64 bits integer to float.

- fcvt.s.l: for converting an signed 64 bits integer to float.

- fcvt.lu.s: for converting a float to unsigned 64 bits integer.

- fcvt.l.s: for converting a float to signed 64 bits integer.

Because of this reason, at the instruction selection phase the
compiler has to provide a custom lowering to instructions that are
actually supported.

Overall, this patch fixes about 50 tests out of 999 from the dnnLib
regression suite, but exhaustive unit tests running the generated code
on a simulator should be developed for complementing the base "codegen"
tests provided with this patch.