[SW-11425] Fixed a bug in mask value copies to MR0
Mask values were copied to MR0 assuming they come from MR registers, so
the maskand instruction was always suitable. However the value sometimes
comes from a GPR register, in which case maskand can not be used and
mov.m.x should instead be used.

Fixed the bug and added asserts in the two places where maskand is emitted
to detect this kind of errors ASAP with debug builds.

Also updated the maskreg.cpp CodeGen test, that was the test unmasking
the bug.