[SW-11425] Enable mask registers in clobber statement for inline asm
The mask register names were not accepted in clobber statements,
but we had some dnnLib code actually clobbering mask registers and
needing to declare that behaviour. This patch enables the feature.

The actual register names are defined in TableGen, but the backend
needs to programatically provide the list of registers for GCC-like
clobbers within its TargetInfo derived class.

As a simple test we add a clobber statement to one of the existing
maskreg.cpp tests cases.