[SW-16747] Fixed the casting between 64 bits integers and floats
The newly added feature for casting from uint64_t and int64_t to float
was untested and did not work. Some failing ArgMax tests from the
dnnLib regression were actually exercising the feature and allowed me
to debug the issues with the newly added code (see SW-17049).

The main problem with the casting feature was that my lowering relied
on a mixture of machine-independent and machine-specific operators.
At the lowering stage one should on product  machine nodes. I was,
for example, creating constants that needed to be materialized on the
DAGToDag stage, which happens before Lowering.

Now all the replacement nodes generated at the Lowering stage are
machine-specific, as expected by the RISC-V backend.

This patch alone brings down the number of test failures on the dnnLib
regression from 259 to just 20 (almost clean).