[SW-16747] Rounding towards zero when casting from float to (u)int32_t
The rounding behaviour when casting from float to (u)int32_t was set
as "dynamic" (using the current value on the FRM CSR), however, rounding
towards zero is what the C99 standard demands and what the backend
should do.

This patch fixes the failures on ResizeBilinearInst and ResizeNearest
tests reported on SW-17089 (12 failures in total).