[SW-16747] Fixed FMADD_S results on casting from (u)int64_t to FP
The FMADD_S instructions produced for lowering ISelDAGToDAG casting from
(u)int64_t to floating point were incorrectly marked as returning four
results, when the instruction just returns one result and it is marked
as such on its TableGen definition.

There was no effect on code generation due to this inconsistency, but
an assert was failing on debug builds.