[SW-16747] The sw-16747.c test needed updating with lastest fixes
The test was still expecting incorrect code and needed to be updated.

More precisely it was expecting the compiler to generate the following
incorrect instruction:

fmadd.s ft0, ft0, 1333788672, ft1, dyn

Note that the third operand is an immediate, but should rather be a
floating point register. The compiler no longer generates that faulty
code and the test should not expect it either.