[SW-11366] The splatAllOnes.ll test needed to be updated
The splatAllOnes.ll test expected the generated code to contain the
following instruction:

fbci.pi ft0, -1

Note however that the second operand for fbci.pi should be a 20 bits
unsigned integer, and the value -1 does not meet that criteria. The fix
for this issue was merged with 1f7b9ec6b73f60ae268923edecefe55a5dceba13.

I did not realize that this test existed, so did not update it as part
of the fix. However, the code generated is now correct, and the test
gets updated with this patch.