[SW-18580] Fixed missing instruction selection for v8i8 stores
A bug was reported affecting the compilation of et-common-libs. Initial
investigation allowed to reduce the reproduction case to the following
small C program:

// clang-11 -std=gnu11 --target=riscv64-unknown-elf -mabi=lp64f -mcpu=et-soc1-min test.c -c -O2
void f(char *buf) {
  unsigned int len = 0;
  while (len < 16) {
    buf[len++] = '0';
  }
}

Compilation with clang aborted due to not having a defined instruction
selection for a v8i8 store operaation.

A simpler way to reproduce the issue not involving clang follows:

// llc -mcpu=et-soc1-min -target-abi lp64f < test.ll
target triple = "riscv64-unknown-unknown-elf"
define void @f16(i8* nocapture %0) local_unnamed_addr #0 {
  %2 = bitcast i8* %0 to <16 x i8>*
  store <16 x i8> <i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48, i8 48>, <16 x i8>* %2, align 1
  ret void
}

The bug happened because there was no isel rule:

- not in RISC-V
- not in Esperanto's TableGen
- not in Esperanto's DAGToDAG rewrite passes
- neither in Esperanto's RISCVISelDAGToDAG customizations

A code reviewed showed that the missing ISel would indeed be done in
Esperanto's RISCVISelDAGToDAG customizatios if it was not because the
logic in esperantoMemop did bail when the address space was not zero.

The reality is that there are cases like when dealing with a vector,
that need the custom rewriting done in esperantoMemop, irrespective of
the address space.

With this patch the rewrite happens for those missing cases, like v8i8.