[SW-19530] Removed check-on-top-of-master from CI
The check-on-top-of-master job was failing with timeouts. The job
was taking too long to git fetch and was not essential for this CI.

If a merge is not on top of master there are two possibilities:

- the merge has no conflict and it is unlikely to introduce any issue
  if it passes CI.

- there is a merge conflict and you will not be able to merge.

Checking whether on top of master only protects us from the possibility
that an automatic merge is wrong and leads to a broken tree in
esperanto/master, even thought the changes passed CI before merge.

I have been doing software for quite a few years and that kind of issue
only happened once to me because of indirect includes. If it ever
happens on esperanto-llvm we will find out inmediately because there
is another CI on esperanto/master after merge. This
check-on-top-of-master does not pay off for the support time that has
consumed so far.